requirements.txt
# Core Data Science & Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.13.0

# Financial Analysis & Technical Indicators
yfinance>=0.2.0
pynance>=1.4.0
TA-Lib>=0.4.24
# Note: TA-Lib might need system-level installation first:
# On Ubuntu: sudo apt-get install ta-lib
# On Mac: brew install ta-lib

# Natural Language Processing & Sentiment Analysis
nltk>=3.8.0
textblob>=0.17.1
vaderSentiment>=3.3.2
spacy>=3.5.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl

# Text Processing & Topic Modeling
gensim>=4.3.0
wordcloud>=1.9.0
transformers>=4.21.0
torch>=2.0.0

# Web Scraping & Data Collection (if needed)
requests>=2.28.0
beautifulsoup4>=4.12.0
selenium>=4.8.0

# Date/Time Handling
python-dateutil>=2.8.0
pytz>=2022.7.0

# Statistical Analysis
statsmodels>=0.13.0

# Configuration & Logging
PyYAML>=6.0
python-dotenv>=1.0.0

# Jupyter & Development
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.0.0
ipywidgets>=8.0.0

# Testing & Code Quality
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# API & Web Framework (for potential dashboard)
fastapi>=0.95.0
uvicorn>=0.21.0
streamlit>=1.22.0

# Utilities
tqdm>=4.65.0
pathlib2>=2.3.0; python_version < '3.4'